Removed interactionComponent2 and instead updated the sensor componet
